@@ -1,48 +1,46 @@
 {
-  "name": "amarisoft-ui",
+  "name": "amarisoftui",
   "private": true,
-  "version": "0.0.1",
-  "description": "Amarisoft UI",
+  "version": "1.0.0",
+  "description": "AmarisoftUI",
   "repository": "",
   "license": "UNLICENSED",
   "scripts": {
     "start": "npm run dev",
     "dev": "cross-env NODE_ENV=development vite",
-    "build": "cross-env NODE_ENV=production vite build",
-    "postinstall": "cpy ./node_modules/framework7-icons/fonts/*.* ./src/fonts/ && cpy ./node_modules/material-icons/iconfont/*.* ./src/fonts/"
+    "build": "cross-env NODE_ENV=production vite build && npx workbox generateSW workbox-config.js",
+    "postinstall": "cpy --flat ./node_modules/framework7-icons/fonts/*.* ./src/fonts/ && cpy --flat ./node_modules/material-icons/iconfont/*.* ./src/fonts/"
   },
   "browserslist": [
-    "Android >= 7",
-    "IOS >= 11",
-    "Safari >= 11",
-    "Chrome >= 49",
-    "Firefox >= 31",
-    "Samsung >= 5"
+    "IOS >= 13",
+    "Safari >= 13",
+    "last 5 Chrome versions",
+    "last 5 Firefox versions",
+    "Samsung >= 12"
   ],
   "dependencies": {
-    "dom7": "^3.0.0",
-    "framework7": "^6.3.14",
+    "dom7": "^4.0.4",
+    "framework7": "^7.0.8",
     "framework7-icons": "^5.0.5",
-    "framework7-react": "^6.3.14",
+    "framework7-react": "^7.0.8",
     "is-ip": "^5.0.0",
     "lazylist-react": "^2.4.0-alpha.0.2",
-    "material-icons": "^1.10.5",
+    "material-icons": "^1.12.1",
     "prop-types": "^15.8.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-json-view": "^1.21.3",
-    "react-singleton-hook": "^3.2.3",
-    "react-use-websocket": "^2.9.1",
-    "signal-js": "^3.0.1",
-    "skeleton-elements": "^3.5.0",
-    "swiper": "^6.8.4"
+    "react-use-websocket": "^4.2.0",
+    "skeleton-elements": "^4.0.1",
+    "swiper": "^8.4.4"
   },
   "devDependencies": {
-    "@peterbee/react-singleton": "github:peterbee/react-singleton",
     "@vitejs/plugin-react-refresh": "^1.3.6",
-    "cpy-cli": "^4.1.0",
+    "cpy-cli": "^4.2.0",
     "cross-env": "^7.0.3",
-    "postcss-preset-env": "^6.7.0",
-    "vite": "^2.7.10"
+    "less": "^4.1.3",
+    "postcss-preset-env": "^7.8.3",
+    "vite": "^3.2.3",
+    "workbox-cli": "^6.5.4"
   }
 }
