@@ -1,6 +1,6 @@
 // Import React and ReactDOM
 import React from 'react';
-import ReactDOM from 'react-dom';
+import { createRoot } from 'react-dom/client';
 
 // Import Framework7
 import Framework7 from 'framework7/lite-bundle';
@@ -9,11 +9,11 @@ import Framework7 from 'framework7/lite-bundle';
 import Framework7React from 'framework7-react';
 
 // Import Framework7 Styles
-import 'framework7/framework7-bundle.css';
+import 'framework7/css/bundle';
 
 // Import Icons and App Custom Styles
 import '../css/icons.css';
-import '../css/app.css';
+import '../css/app.less';
 
 // Import App Component
 import App from '../components/app.jsx';
@@ -22,7 +22,5 @@ import App from '../components/app.jsx';
 Framework7.use(Framework7React)
 
 // Mount React App
-ReactDOM.render(
-  React.createElement(App),
-  document.getElementById('app'),
-);
+const root = createRoot(document.getElementById('app'));
+root.render(React.createElement(App));
